Update README.md
I found it very difficult to install k_aug so that I could use it in my trust-region filter algorithm. After 2 months of effort, I finally managed to install it (and it worked), and I have composed a detailed manual for installing this sensitivity package on the latest edition of Windows.